+++ src/pass/licm.rs
@@ -1,0 +1,155 @@
+// Copyright (c) 2017-2019 Fabian Schuiki
+
+//! Loop Independent Code Motion
+
+use crate::ir::prelude::*;
+use crate::opt::prelude::*;
+use crate::pass::gcse::{DominatorTree, PredecessorTable};
+use std::collections::{HashMap, HashSet};
+
+/// Loop Independent Code Motion
+///
+/// This moves all instructions as far upwards in the control flow graph as
+/// possible given the point of declaration of their arguments.
+pub struct LoopIndepCodeMotion;
+
+impl Pass for LoopIndepCodeMotion {
+    fn run_on_cfg(ctx: &PassContext, unit: &mut impl UnitBuilder) -> bool {
+        info!("LICM [{}]", unit.unit().name());
+        let mut modified = false;
+
+        // Build the predecessor table and dominator tree.
+        let pred = PredecessorTable::new(unit.dfg(), unit.func_layout());
+        let dt = DominatorTree::new(unit.func_layout(), &pred);
+
+        // Create a work queue which allows us to process the blocks in control
+        // flow order. Also number the blocks as we go.
+        let mut block_numbers = HashMap::<Block, usize>::new();
+        let mut work_done = HashSet::<Block>::new();
+        let mut work_pending = HashSet::<Block>::new();
+        let entry = unit.func_layout().entry();
+        work_pending.insert(entry);
+        block_numbers.insert(entry, 0);
+
+        while let Some(&block) = work_pending.iter().next() {
+            work_pending.remove(&block);
+            work_done.insert(block);
+            trace!("Working on {}", block.dump(unit.cfg()));
+
+            // Process the instructions in this block.
+            for inst in unit.func_layout().insts(block).collect::<Vec<_>>() {
+                modified |= move_instruction(ctx, unit, block, inst, &dt, &block_numbers);
+            }
+
+            // Work on the successors of this block.
+            let term = unit.func_layout().terminator(block);
+            if unit.dfg()[term].opcode().is_terminator() {
+                work_pending.extend(
+                    unit.dfg()[term]
+                        .blocks()
+                        .iter()
+                        .cloned()
+                        .filter(|bb| !work_done.contains(bb)),
+                );
+                let next_number = block_numbers[&block] + 1;
+                for bb in unit.dfg()[term].blocks().iter().cloned() {
+                    block_numbers.entry(bb).or_insert(next_number);
+                }
+            }
+        }
+
+        trace!("Final block numbers:");
+        for (bb, num) in block_numbers {
+            trace!("  {} = {}", bb.dump(unit.cfg()), num);
+        }
+
+        modified
+    }
+}
+
+fn move_instruction(
+    _ctx: &PassContext,
+    unit: &mut impl UnitBuilder,
+    block: Block,
+    inst: Inst,
+    dt: &DominatorTree,
+    block_numbers: &HashMap<Block, usize>,
+) -> bool {
+    // Some instructions we don't want to touch.
+    let op = unit.dfg()[inst].opcode();
+    if op == Opcode::Ld
+        || op == Opcode::St
+        || op == Opcode::Prb
+        || op == Opcode::Drv
+        || op == Opcode::Phi
+        || op.is_terminator()
+    {
+        return false;
+    }
+    trace!("  Working on {}", inst.dump(unit.dfg(), unit.try_cfg()));
+
+    // To determine the possible insertion locations, we first need to find for
+    // each argument of this instruction, which blocks its definition dominates.
+    let doms: Vec<_> = unit.dfg()[inst]
+        .args()
+        .iter()
+        .flat_map(|&arg| unit.dfg().get_value_inst(arg))
+        .map(|inst| unit.func_layout().inst_block(inst).unwrap())
+        .map(|block| dt.dominated_by(block))
+        .collect();
+
+    // If the instruction depends on nothing (e.g. constants), move them up into
+    // the entry block.
+    if doms.is_empty() {
+        let layout = unit.func_layout_mut();
+        let entry = layout.entry();
+        let entry_term = layout.terminator(entry);
+        layout.remove_inst(inst);
+        layout.insert_inst_before(inst, entry_term);
+        debug!(
+            "Move {} into {}",
+            inst.dump(unit.dfg(), unit.try_cfg()),
+            entry.dump(unit.cfg())
+        );
+        return true;
+    }
+    // trace!("    Dominated blocks: {:?}", doms);
+
+    // Find the blocks that are present in all the domination sets. These are
+    // the blocks where the current instruction can safely be moved and still
+    // "see" all its arguments.
+    let possible_bbs = doms
+        .get(0)
+        .into_iter()
+        .flat_map(|bbs| bbs.iter())
+        .filter(|bb| doms.iter().all(|bbs| bbs.contains(bb)))
+        .cloned();
+
+    // Decide on the best block to move the instruction to. This is given as the
+    // block with the lowest number, which is the lowest number of control flow
+    // edges from the entry block.
+    let best_bb = possible_bbs
+        .flat_map(|bb| block_numbers.get(&bb).map(|&num| (bb, num)))
+        .min_by_key(|&(_, num)| num)
+        .map(|(bb, _)| bb);
+    let best_bb = match best_bb {
+        Some(bb) => bb,
+        None => return false,
+    };
+    trace!("    Best block: {}", best_bb.dump(unit.cfg()));
+
+    // Move the instruction up into this block.
+    if best_bb == block {
+        return false;
+    }
+    debug!(
+        "Move {} into {}",
+        inst.dump(unit.dfg(), unit.try_cfg()),
+        best_bb.dump(unit.cfg())
+    );
+    let layout = unit.func_layout_mut();
+    let term = layout.terminator(best_bb);
+    layout.remove_inst(inst);
+    layout.insert_inst_before(inst, term);
+    true
+}