+++ src/pass/mod.rs
@@ -8,9 +8,11 @@
 pub mod cf;
 pub mod dce;
 pub mod gcse;
+pub mod licm;
 pub mod tcm;
 
 pub use cf::ConstFolding;
 pub use dce::DeadCodeElim;
 pub use gcse::GlobalCommonSubexprElim;
+pub use licm::LoopIndepCodeMotion;
 pub use tcm::TemporalCodeMotion;