+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Global Common Subexpression Elimination pass.
 - Add Temporal Code Motion pass.
 - Add `phi` instruction.
+- Add Loop Independent Code Motion Pass.
 
 ### Changed
 - Add missing `dyn` keywords.